+++ src/Commands/Commands.cpp
@@ -14,6 +14,7 @@
 #include "ToggleMessageList.h"
 #include "DeselectObject.h"
 #include "DeselectObject5.h"
+#include "ToggleCameos.h"
 
 #include <CCINIClass.h>
 
@@ -32,6 +33,7 @@ DEFINE_HOOK(0x533066, CommandClassCallback_Register, 0x6)
 	MakeCommand<ToggleSWSidebar>();
 	MakeCommand<DeselectObjectCommandClass>();
 	MakeCommand<DeselectObject5CommandClass>();
+	MakeCommand<ToggleCameosCommandClass>();
 
 	if (Phobos::Config::SuperWeaponSidebarCommands)
 	{