+++ src/Ext/TechnoType/Body.cpp
@@ -1344,7 +1344,11 @@ void TechnoTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 
 	this->ShadowIndex_Frame.Read(exArtINI, pArtSection, "ShadowIndex.Frame");
 
+	this->CameoPCX.Read(pArtINI, pArtSection, "CameoPCX");
 	this->AltCameoPCX.Read(pArtINI, pArtSection, "AltCameoPCX");
+	this->ShowCameo_CustomPalette.LoadFromINI(pArtINI, pArtSection, "ShowCameo.CustomPalette");
+	this->ShowCameo_CustomShape.Read(exArtINI, pArtSection, "ShowCameo.CustomShape");
+	this->ShowCameo_CustomPCX.Read(pArtINI, pArtSection, "ShowCameo.CustomPCX");
 
 	this->LaserTrailData.clear();
 	for (size_t i = 0; ; ++i)
@@ -1464,6 +1468,15 @@ void TechnoTypeExt::ExtData::LoadFromINIFile(CCINIClass* const pINI)
 	// VoiceIFVRepair from Ares 0.2
 	this->VoiceIFVRepair.Read(exINI, pSection, "VoiceIFVRepair");
 	this->ParseVoiceWeaponAttacks(exINI, pSection, this->VoiceWeaponAttacks, this->VoiceEliteWeaponAttacks);
+
+	this->ShowCameo.Read(exINI, pSection, "ShowCameo");
+	this->ShowCameo_PerRow.Read(exINI, pSection, "ShowCameo.PerRow");
+	this->ShowCameo_BottomOffset.Read(exINI, pSection, "ShowCameo.BottomOffset");
+	this->ShowCameo_Translucency.Read(exINI, pSection, "ShowCameo.Translucency");
+	this->ShowCameo_OverlapXY.Read(exINI, pSection, "ShowCameo.OverlapXY");
+	this->ShowCameo_OffsetXY.Read(exINI, pSection, "ShowCameo.OffsetXY");
+	this->ShowCameo_OverlapPrimary.Read(exINI, pSection, "ShowCameo.OverlapPrimary");
+	this->CameoPalette.LoadFromINI(pArtINI, pArtSection, "CameoPalette");
 }
 
 void TechnoTypeExt::ExtData::LoadFromINIByWhatAmI(INI_EX& exINI, const char* pSection, INI_EX& exArtINI, const char* pArtSection)
@@ -1533,7 +1546,11 @@ void TechnoTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->ImmuneToCrit)
 		.Process(this->MultiMindControl_ReleaseVictim)
 		.Process(this->CameoPriority)
+		.Process(this->CameoPCX)
 		.Process(this->AltCameoPCX)
+		.Process(this->ShowCameo_CustomPalette)
+		.Process(this->ShowCameo_CustomShape)
+		.Process(this->ShowCameo_CustomPCX)
 		.Process(this->NoManualMove)
 		.Process(this->InitialStrength)
 		.Process(this->ReloadInTransport)
@@ -1975,6 +1992,18 @@ void TechnoTypeExt::ExtData::Serialize(T& Stm)
 		.Process(this->Missile_Cruise)
 		.Process(this->Missile_TakeOffAnim)
 		.Process(this->Missile_TakeOffSeparation)
+
+		.Process(this->ShowCameo)
+		.Process(this->ShowCameo_PerRow)
+		.Process(this->ShowCameo_BottomOffset)
+		.Process(this->ShowCameo_Translucency)
+		.Process(this->ShowCameo_OverlapXY)
+		.Process(this->ShowCameo_OffsetXY)
+		.Process(this->ShowCameo_OverlapPrimary)
+		.Process(this->CameoPalette)
+		.Process(this->ShowCameo_CustomPalette)
+		.Process(this->ShowCameo_CustomShape)
+		.Process(this->ShowCameo_CustomPCX)
 		;
 }
 void TechnoTypeExt::ExtData::LoadFromStream(PhobosStreamReader& Stm)